pyproject.toml
[project]
name = "peek-kit"
version = "0.2.0"
description = "AI-Powered App Analysis & Audit as a Claude Code MCP"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Ufuk Karaca", email = "ufuk@rodyr.com"}
]
dependencies = [
    "mcp[cli]>=1.3",
    "pyobjc-framework-ApplicationServices>=10.0",
    "pyobjc-framework-Quartz>=10.0",
    "pyobjc-framework-Cocoa>=10.0",
    "atomacos>=3.0",
    "Pillow>=10.0",
    "pydantic>=2.0",
    "rich>=13.0",
    "jinja2>=3.1",
]

[project.urls]
Homepage = "https://github.com/ufukkaraca/peek-kit"
Repository = "https://github.com/ufukkaraca/peek-kit.git"

[project.scripts]
peek-kit = "peek_kit.server:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
